% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_prob_mmcd}
\alias{clean_prob_mmcd}
\title{Probability of obtaining at least one clean h-subset in the \code{\link{mmcd}} function.}
\usage{
clean_prob_mmcd(p, q, n_subsets = 500, contamination = 0.5)
}
\arguments{
\item{p}{number of rows.}

\item{q}{number of columns.}

\item{n_subsets}{number of elemental h-substs (default is 500).}

\item{contamination}{level of contamination (default is 0.5).}
}
\value{
Probability of obtaining at least one clean h-subset in the \code{\link{mmcd}} function.
}
\description{
Probability of obtaining at least one clean h-subset in the \code{\link{mmcd}} function.
}
