% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmcd}
\alias{mmcd}
\title{The Matrix Minimum Covariance Determinant (MMCD) Estimator}
\usage{
mmcd(
  X,
  nsamp = 500L,
  alpha = 0.5,
  cov_row_init = NULL,
  cov_col_init = NULL,
  diag = "none",
  lambda = 0,
  max_iter_cstep = 100L,
  max_iter_MLE = 100L,
  max_iter_cstep_init = 2L,
  max_iter_MLE_init = 2L,
  nmini = 300L,
  nsub_pop = 1500L,
  adapt_alpha = TRUE,
  reweight = TRUE,
  scale_consistency = "quant",
  outlier_quant = 0.975,
  nthreads = 1L
)
}
\arguments{
\item{X}{a 3d array of dimension \eqn{(p,q,n)}, containing \eqn{n} matrix-variate samples of \eqn{p} rows and \eqn{q} columns in each slice.}

\item{nsamp}{number of initial h-subsets (default is 500).}

\item{alpha}{numeric parameter between 0.5 (default) and 1. Controls the size \eqn{h \approx alpha * n} of the h-subset over which the determinant is minimized.}

\item{cov_row_init}{matrix. Initial \code{cov_row} \eqn{p \times p} matrix. Identity by default.}

\item{cov_col_init}{matrix. Initial \code{cov_col} \eqn{p \times p} matrix. Identity by default.}

\item{diag}{Character. If "none" (default) all entries of \code{cov_row} and \code{cov_col} are estimated. If either "both", "row", or "col" only the diagonal entries of the respective matrices are estimated.}

\item{lambda}{a smooting parameter for the rowwise and columnwise covariance matrices.}

\item{max_iter_cstep}{upper limit of C-step iterations (default is 100)}

\item{max_iter_MLE}{upper limit of MLE iterations (default is 100)}

\item{max_iter_cstep_init}{upper limit of C-step iterations for initial h-subsets (default is 2)}

\item{max_iter_MLE_init}{upper limit of MLE iterations for initial h-subsets (default is 2)}

\item{nmini}{for \eqn{n > 2*nmini} the algorithm splits the data into smaller subsets with at least \code{nmini} samples for initialization (default is 300)}

\item{nsub_pop}{for \eqn{n > 2*nmini} a subpopulation of \eqn{\min(n,nsub\_pop)} samples is divided into \eqn{k = floor(\min(n,nsub\_pop)/nmini)} subsets  for initialization (default is 1500)}

\item{adapt_alpha}{Logical. If TRUE (default) alpha is adapted to take the dimension of the data into account.}

\item{reweight}{Logical. If TRUE (default) the reweighted MMCD estimators are computed.}

\item{scale_consistency}{Character. Either "quant" (default) or "mmd_med". If "quant", the consistency factor is chosen to achieve consistency under the matrix normal distribution.
If "mmd_med", the consistency factor is chosen based on the Mahalanobis distances of the observations.}

\item{outlier_quant}{numeric parameter between 0 and 1. Chi-square quantile used in the reweighting step.}

\item{nthreads}{Integer. If 1 (default), all computations are carried out sequentially.
If larger then 1, C-steps are carried out in parallel using \code{nthreads} threads.
If < 0, all possible threads are used.}
}
\value{
A list containing the following:
\item{\code{mu}}{Estimated \eqn{p \times q} mean matrix.}
\item{\code{cov_row}}{Estimated \eqn{p} times \eqn{p} rowwise covariance matrix.}
\item{\code{cov_col}}{Estimated \eqn{q} times \eqn{q} columnwise covariance matrix.}
\item{\code{cov_row_inv}}{Inverse of \code{cov_row}.}
\item{\code{cov_col_inv}}{Inverse of \code{cov_col}.}
\item{\code{md}}{Squared Mahalanobis distances.}
\item{\code{md_raw}}{Squared Mahalanobis distances based on \emph{raw} MMCD estimators.}
\item{\code{det}}{Value of objective function (determinant of Kronecker product of rowwise and columnwise covariane).}
\item{\code{alpha}}{The (adjusted) value of alpha used to determine the size of the h-subset.}
\item{\code{consistency_factors}}{Consistency factors for raw and reweighted MMCD estimators.}
\item{\code{dets}}{Objective values for the final h-subsets.}
\item{\code{best_i}}{ID of subset with best objective.}
\item{\code{h_subset}}{Final h-subset of \emph{raw} MMCD estimators.}
\item{\code{h_subset_reweighted}}{Final h-subset of \emph{reweighted} MMCD estimators.}
\item{\code{iterations}}{Number of C-steps.}
\item{\code{dets_init_first}}{Objective values for the \code{nsamp} initial h-subsets after \code{max_iter_cstep_init} C-steps.}
\item{\code{subsets_first}}{Subsets created in subsampling procedure for large \code{n}.}
\item{\code{dets_init_second}}{Objective values of the 10 best initial subsets after executing C-steps until convergence.}
}
\description{
\code{mmcd} computes the robust MMCD estimators of location and covariance for matrix-variate data
using the FastMMCD algorithm \insertCite{mayrhofer2024}{robustmatrix}.
}
\details{
The MMCD estimators generalize the well-known Minimum Covariance Determinant (MCD)
\insertCite{Rousseeuw1985,Rousseeuw1999}{robustmatrix} to the matrix-variate setting.
It looks for the \eqn{h} observations, \eqn{h = \alpha * n}, whose covariance matrix has the smallest determinant.
The FastMMCD algorithm is used for computation and is described in detail in \insertCite{mayrhofer2024}{robustmatrix}.
NOTE: The procedure depends on \emph{random} initial subsets. Currently setting a seed is only possible if \code{nthreads = 1}.
}
\examples{
n = 1000; p = 2; q = 3
mu = matrix(rep(0, p*q), nrow = p, ncol = q)
cov_row = matrix(c(1,0.5,0.5,1), nrow = p, ncol = p)
cov_col = matrix(c(3,2,1,2,3,2,1,2,3), nrow = q, ncol = q)
X <- rmatnorm(n = n, mu, cov_row, cov_col)
ind <- sample(1:n, 0.3*n)
X[,,ind] <- rmatnorm(n = length(ind), matrix(rep(10, p*q), nrow = p, ncol = q), cov_row, cov_col)
par_mmle <- mmle(X)
par_mmcd <- mmcd(X)
distances_mmle <- mmd(X, par_mmle$mu, par_mmle$cov_row, par_mmle$cov_col)
distances_mmcd <- mmd(X, par_mmcd$mu, par_mmcd$cov_row, par_mmcd$cov_col)
plot(distances_mmle, distances_mmcd)
abline(h = qchisq(0.99, p*q), lty = 2, col = "red")
abline(v = qchisq(0.99, p*q), lty = 2, col = "red")
}
\references{
\insertAllCited{}
}
\seealso{
The \code{mmcd} algorithm uses the \code{\link{cstep}} and \code{\link{mmle}} functions.
}
