% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{n_subsets_mmcd}
\alias{n_subsets_mmcd}
\title{Number of subsets that are required to obtain at least one clean h-subset in the \code{\link{mmcd}} function with probability \code{prob}.}
\usage{
n_subsets_mmcd(p, q, prob = 0.99, contamination = 0.5)
}
\arguments{
\item{p}{number of rows.}

\item{q}{number of columns.}

\item{prob}{probability (default is 0.99).}

\item{contamination}{level of contamination (default is 0.5).}
}
\value{
Number of subsets that are required to obtain at least one clean h-subset in the \code{\link{mmcd}} function with probability \code{prob}.
}
\description{
Number of subsets that are required to obtain at least one clean h-subset in the \code{\link{mmcd}} function with probability \code{prob}.
}
